src/requirements.txt
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-bootstrap3==11.0.0
django-braces==1.13.0
django-countries==6.1.3
django-debug-toolbar==1.10.1
django-formtools==2.1
django-resized==0.3.11
django-tinymce==2.7.0
djangorestframework==3.9.0
html5lib==1.0.1
idna==2.7
markdown2==2.3.5
oauthlib==2.1.0
pdfcrowd==4.3.5
peewee==3.6.4
Pillow==5.2.0
pycparser==2.18
PyPDF2==1.26.0
Pyphen==0.9.5
python3-openid==3.1.0
pytz==2018.5
reportlab==3.5.9
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.2.4
tinycss2==0.6.1
urllib3==1.24
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
webencodings==0.5.1
